requirements_test.txt
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-socket==0.7.0
pytest-sugar==1.0.0
pytest-timeout==2.2.0
pytest==8.1.0
